# web/openapi.json
{
  "components": {
    "schemas": {
...
      "CoverageStatus": {
        "enum": [
          "exact_covered",
          "fallback_only",
          "mixed",
          "uncovered"
        ],
        "title": "CoverageStatus",
        "type": "string"
      },
      "DeleteCoverageResponse": {
        "additionalProperties": false,
        "properties": {
          "coverage_id": {
            "title": "Coverage Id",
            "type": "string"
          },
          "deleted": {
            "title": "Deleted",
            "type": "boolean"
          }
        },
        "required": [
          "coverage_id",
          "deleted"
        ],
        "title": "DeleteCoverageResponse",
        "type": "object"
      },
...
      "FieldError": {
        "additionalProperties": false,
        "properties": {
          "field": {
            "title": "Field",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "field",
          "message"
        ],
        "title": "FieldError",
        "type": "object"
      },
...
      "OpportunityDrillDownResponse": {
        "additionalProperties": false,
        "properties": {
          "configuration": {
            "$ref": "#/components/schemas/PlanningConfigurationResponse"
          },
          "coverage_status": {
            "$ref": "#/components/schemas/CoverageStatus"
          },
          "model_year_demand": {
            "$ref": "#/components/schemas/ModelYearDemandResponse"
          }
        },
        "required": [
          "configuration",
          "model_year_demand",
          "coverage_status"
        ],
        "title": "OpportunityDrillDownResponse",
        "type": "object"
      },
      "OpportunityGroupBy": {
        "enum": [
          "brand",
          "model",
          "model_year"
        ],
        "title": "OpportunityGroupBy",
        "type": "string"
      },
      "OpportunityPageResponse": {
        "additionalProperties": false,
        "properties": {
          "integrity_warnings": {
            "items": {
              "type": "string"
            },
            "title": "Integrity Warnings",
            "type": "array"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/OpportunityRowResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "pages": {
            "title": "Pages",
            "type": "integer"
          },
          "snapshot_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snapshot Id"
          },
          "strategy_name": {
            "title": "Strategy Name",
            "type": "string"
          },
          "strategy_version": {
            "title": "Strategy Version",
            "type": "string"
          },
          "summary": {
            "$ref": "#/components/schemas/OpportunitySummaryResponse"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "versions": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SnapshotVersionsResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "items",
          "summary",
          "strategy_name",
          "strategy_version",
          "integrity_warnings",
          "total",
          "page",
          "page_size",
          "pages"
        ],
        "title": "OpportunityPageResponse",
        "type": "object"
      },
      "OpportunityRowResponse": {
        "additionalProperties": false,
        "properties": {
          "brand": {
            "title": "Brand",
            "type": "string"
          },
          "contributing_configuration_count": {
            "title": "Contributing Configuration Count",
            "type": "integer"
          },
          "coverage_status": {
            "$ref": "#/components/schemas/CoverageStatus"
          },
          "data_version": {
            "title": "Data Version",
            "type": "string"
          },
          "demand": {
            "$ref": "#/components/schemas/DemandRangeResponse"
          },
          "evidence_status": {
            "$ref": "#/components/schemas/EvidenceStatus"
          },
          "exact_covered_base_units": {
            "title": "Exact Covered Base Units",
            "type": "integer"
          },
          "fallback_covered_base_units": {
            "title": "Fallback Covered Base Units",
            "type": "integer"
          },
          "generation_basis": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generation Basis"
          },
          "generation_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generation Name"
          },
          "group_by": {
            "title": "Group By",
            "type": "string"
          },
          "group_id": {
            "title": "Group Id",
            "type": "string"
          },
          "icor_worked_base_units": {
            "title": "Icor Worked Base Units",
            "type": "integer"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "model_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Year"
          },
          "score": {
            "$ref": "#/components/schemas/OpportunityScoreResponse"
          },
          "uncovered_base_units": {
            "title": "Uncovered Base Units",
            "type": "integer"
          }
        },
        "required": [
          "group_id",
          "group_by",
          "brand",
          "model",
          "model_year",
          "generation_name",
          "generation_basis",
          "icor_worked_base_units",
          "demand",
          "contributing_configuration_count",
          "exact_covered_base_units",
          "fallback_covered_base_units",
          "uncovered_base_units",
          "coverage_status",
          "score",
          "evidence_status",
          "data_version"
        ],
        "title": "OpportunityRowResponse",
        "type": "object"
      },
...
      "OpportunitySummaryResponse": {
        "additionalProperties": false,
        "properties": {
          "base_units": {
            "title": "Base Units",
            "type": "integer"
          },
          "exact_covered_base_units": {
            "title": "Exact Covered Base Units",
            "type": "integer"
          },
          "high_demand_uncovered_base_units": {
            "title": "High Demand Uncovered Base Units",
            "type": "integer"
          }
        },
        "required": [
          "base_units",
          "exact_covered_base_units",
          "high_demand_uncovered_base_units"
        ],
        "title": "OpportunitySummaryResponse",
        "type": "object"
      },
...
      "ProblemResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "correlation_id": {
            "title": "Correlation Id",
            "type": "string"
          },
          "field_errors": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "title": "Field Errors",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "correlation_id"
        ],
        "title": "ProblemResponse",
        "type": "object"
      },
      "ProductionCoverageRequest": {
        "additionalProperties": false,
        "properties": {
          "brand": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand"
          },
          "configuration_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Configuration Id"
          },
          "match_type": {
            "$ref": "#/components/schemas/CoverageMatchType"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "model_year": {
            "title": "Model Year",
            "type": "integer"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "required": [
          "match_type",
          "configuration_id",
          "brand",
          "model",
          "model_year"
        ],
        "title": "ProductionCoverageRequest",
        "type": "object"
      },
      "ProductionCoverageResponse": {
        "additionalProperties": false,
        "properties": {
          "brand": {
            "title": "Brand",
            "type": "string"
          },
          "configuration_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Configuration Id"
          },
          "coverage_id": {
            "title": "Coverage Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "match_type": {
            "$ref": "#/components/schemas/CoverageMatchType"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "model_year": {
            "title": "Model Year",
            "type": "integer"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "sku": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sku"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "coverage_id",
          "match_type",
          "configuration_id",
          "brand",
          "model",
          "model_year",
          "sku",
          "note",
          "created_at",
          "updated_at"
        ],
        "title": "ProductionCoverageResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Local evidence-led contract for vehicle registration analysis and planning.",
    "title": "ICOR Planner API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/opportunities": {
      "get": {
        "operationId": "opportunities_api_v1_opportunities_get",
        "parameters": [
          {
            "in": "query",
            "name": "group_by",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/OpportunityGroupBy",
              "default": "brand"
            }
          },
          {
            "in": "query",
            "name": "market",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Market"
            }
          },
          {
            "in": "query",
            "name": "horizon",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "integer"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Horizon"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpportunityPageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Opportunities"
      }
    },
    "/api/v1/opportunities/{group_id}/configurations": {
      "get": {
        "operationId": "opportunity_configurations_api_v1_opportunities__group_id__configurations_get",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "group_by",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/OpportunityGroupBy",
              "default": "brand"
            }
          },
          {
            "in": "query",
            "name": "market",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Market"
            }
          },
          {
            "in": "query",
            "name": "horizon",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "integer"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Horizon"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OpportunityDrillDownResponse"
                  },
                  "title": "Response Opportunity Configurations Api V1 Opportunities  Group Id  Configurations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Opportunity Configurations"
      }
    },
...
    "/api/v1/production-coverage": {
      "get": {
        "operationId": "production_coverage_api_v1_production_coverage_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProductionCoverageResponse"
                  },
                  "title": "Response Production Coverage Api V1 Production Coverage Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Production Coverage"
      },
      "post": {
        "operationId": "create_production_coverage_api_v1_production_coverage_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductionCoverageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductionCoverageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Create Production Coverage"
      }
    },
    "/api/v1/production-coverage/{coverage_id}": {
      "delete": {
        "operationId": "delete_production_coverage_api_v1_production_coverage__coverage_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "coverage_id",
            "required": true,
            "schema": {
              "title": "Coverage Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteCoverageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Delete Production Coverage"
      },
      "put": {
        "operationId": "update_production_coverage_api_v1_production_coverage__coverage_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "coverage_id",
            "required": true,
            "schema": {
              "title": "Coverage Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductionCoverageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductionCoverageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Update Production Coverage"
      }
    },
...
  }
}